Refactoring and docstrings
removed the get generation data function. With the configurability, we can increase the volume of images to simulate the high CPU scenario. No need to fake it now.
refactored some of the larger, multi task functions to create better readability
added some docstrings on the functions to aid in understanding what is going on without the visual aid of powerpoint