Exclude azure from package build
Similar to https://github.com/Azure/azure-functions-python-library/pull/97

Excluded azure from the packages. This fixes an issue where the `azure.function` and other azure package versions do not get updated when a version is specified in the functions requirements.txt file.

The reason for this is that the azure directory has an `__init__.py` file, including azure in the package build overrides the user's dependencies.